api-service/requirements.txt
fastapi
uvicorn[standard]
boto3
aiokafka
pydantic
python-multipart
pydantic-settings
loguru
httpx
python-docx
pdfminer.six